Moved Market Transition Menu Out of Main
Yes, it took me this long to figure out how enums work. I only cried a little
